requirements.txt
gradio==4.12.0
openai==1.6.1
google-generativeai==0.3.2